chore: move tests to the same test file
